Protocol edits
Prioritize tuning single setting for axial velocity controller for use
in the rest of the tests. Axial velocity target of 1.0 m/s for rest of
tests to be consistent with the open loop steady state the 0.5 control
input controller was designed around. Suggestion to reduce runtime of
turn test in case there’s not enough testing space/the beluga will go
too far in 60 seconds at 1.0 m/s.